support for GPT generated state machines and lambda functions

+++ src/commands/generate/generate.js
@@ -4,12 +4,14 @@ const { Configuration, OpenAIApi } = require("openai");
 const settingsUtil = require("../../shared/settingsUtil");
 const fs = require("fs-extra");
 var Spinner = require('cli-spinner').Spinner;
+const baseFile = require("../../shared/baseFile.json");
 
 var spinner = new Spinner('Waiting for ChatGPT... %s');
 spinner.setSpinnerString('|/-\\');
 async function run(cmd) {
   console.log("*** Note: This is an experimental feature and depends on the ChatGPT API. Make sure you review the output carefully before using it in production ***")
-  if (cmd.output.toLowerCase() === "cdk" && !cmd.outputFile) {
+  const output = cmd.output.toLowerCase();
+  if (output !== "sam" && !cmd.outputFile) {
     console.log(`You need to specify an output file with --output-file`);
     return;
   }
@@ -20,27 +22,52 @@ async function run(cmd) {
     );
     return;
   }
-  if (!fs.existsSync(cmd.template)) {
-    console.log(
-      `Can't find ${cmd.template}. Use -t option to specify template filename`
-    );
-    const create = await inputUtil.prompt(`Create ${cmd.template}?`);
-    if (create) {
-      fs.writeFileSync(cmd.template, parser.stringify("yaml", baseFile));
-    } else {
-      return;
+  let ownTemplate;
+  if (output === "sam") {
+    if (!fs.existsSync(cmd.template)) {
+      console.log(
+        `Can't find ${cmd.template}. Use -t option to specify template filename`
+      );
+      const create = await inputUtil.prompt(`Create ${cmd.template}?`);
+      if (create) {
+        fs.writeFileSync(cmd.template, parser.stringify("yaml", baseFile));
+      } else {
+        return;
+      }
     }
+
+    ownTemplate = parser.parse("own", fs.readFileSync(cmd.template));
+    ownTemplate.Resources = ownTemplate.Resources || {};
   }
-
-  const ownTemplate = parser.parse("own", fs.readFileSync(cmd.template));
-  ownTemplate.Resources = ownTemplate.Resources || {};
-
   const configuration = new Configuration({
     apiKey,
   });
   const openai = new OpenAIApi(configuration);
-  const format = cmd.output.toLowerCase() === "sam" ? "SAM JSON" : "TypeScript CDK";
-  const prompt = `Generate this in AWS ${format}: ${cmd.query}`;
+  let outputString;
+  let language;
+  switch (output.substring(0, 3)) {
+    case "sam":
+      outputString = "SAM JSON";
+      break;
+    case "cdk":
+      outputString = "TypeScript CDK";
+      break;
+    case "lam":
+      outputString = "Lambda";
+      language = output.split("-")[1];
+      if (language) {
+        outputString += ` in ${language}`;
+      }
+
+      break;
+    case "asl":
+      outputString = "StepFunctions ASL YAML";
+      break;
+    default:
+      console.log(`Invalid output format ${output}. Valid values are 'SAM', 'CDK', 'lambda-<language>' or 'ASL'`);
+      return;
+  }
+  const prompt = `Generate this in AWS ${outputString}: ${cmd.query}. Only return code.`;
 
   const openAiRequest = {
     model: cmd.model,
@@ -56,10 +83,10 @@ async function run(cmd) {
   spinner.start();
   const response = await openai.createChatCompletion(openAiRequest);
   spinner.stop();
-  const text = response.data.choices[0].message.content;
+  let text = response.data.choices[0].message.content;
   // get the first JSON object in the text
   let obj;
-  if (cmd.output.toLowerCase() === "sam") {
+  if (output === "sam") {
     try {
       obj = JSON.parse(text.replace(/\n/g, '').replace(/```/g, '').match(/{.*}/)[0]);
     } catch (e) {
@@ -76,14 +103,23 @@ async function run(cmd) {
 
     console.log(`\n\nGenerated the following resources:\n\n${parser.stringify("yaml", obj)}`);
   } else {
-    console.log(`\n\nGenerated the following CDK code:\n\n${text}`);
+    console.log(text);
+    //check if text has a row starting with ``` followed by text. If so, remove that text
+    const match = text.match(/```.+\n/);
+    if (match && match[0]) {
+      text = text.replace(match[0], '```\n');
+    }
+    if (text.match(/```/g)) {
+      text = text.replace(/\n/g, "¶").match(/```.*```/)[0].split("¶").join('\n').replace(/```/g, '');
+    }
+    console.log(`\n\nGenerated the following ${language} code:\n\n${text}`);
   }
-  const cont = await inputUtil.prompt(cmd.output.toLowerCase() === "sam" ? `Add to template?` : `Add to ${cmd.outputFile}?`);
+  const cont = await inputUtil.prompt(output === "sam" ? `Add to template?` : `Add to ${cmd.outputFile}?`);
   if (!cont) {
     return;
   }
 
-  if (cmd.output.toLowerCase() === "cdk") {
+  if (output !== "sam") {
     if (!cmd.outputFile) {
       console.log(`You need to specify an output file with --output-file`);
       return;
